data size from int to size_t

+++ include/bst.h
@@ -27,7 +27,7 @@ typedef struct Bst_node
 
 typedef struct Bst_iterator
 {
-    int         ____size_of;
+    size_t      ____size_of;
     Bst_node    *____node;
 }Bst_iterator;
 
@@ -35,7 +35,7 @@ typedef struct Bst
 {
     Bst_node    *____root;
     size_t      ____nodes;
-    int         ____size_of;
+    size_t      ____size_of;
 
     int (*____cmp)(void* a,void *b);
 }Bst;
